알림 서버 분리
