+++ .idea/modules.xml
@@ -4,9 +4,8 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/.idea/B-Dool-Server.iml" filepath="$PROJECT_DIR$/.idea/B-Dool-Server.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/chatbot-service.main.iml" filepath="$PROJECT_DIR$/.idea/modules/chatbot-service.main.iml" />
-      <module fileurl="file://$PROJECT_DIR$/auth-service/auth-service.iml" filepath="$PROJECT_DIR$/auth-service/auth-service.iml" />
-      <module fileurl="file://$PROJECT_DIR$/.idea/modules/chatbot-service.main.iml" filepath="$PROJECT_DIR$/.idea/modules/chatbot-service.main.iml" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/member-hub-service.main.iml" filepath="$PROJECT_DIR$/.idea/modules/member-hub-service.main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/notification-service/notification-service.iml" filepath="$PROJECT_DIR$/notification-service/notification-service.iml" />
     </modules>
   </component>
 </project>
