--- member-hub-service/src/main/java/com/bdool/memberhubservice/notification/domain/notification/entity/NotificationType.java
@@ -1,8 +1,0 @@
-package com.bdool.memberhubservice.notification.domain.notification.entity;
-
-public enum NotificationType {
-    WORKSPACE_ENTRY,  // 워크스페이스 입장 알림
-    CALENDAR_INVITE,  // 일정 초대 알림
-    CHANNEL_MESSAGE,  // 채널 메시지 알림
-    SYSTEM_ALERT      // 시스템 알림
-}
