--- member-hub-service/src/main/java/com/bdool/memberhubservice/notification/domain/setting/controller/NotificationTargetSettingController.java
@@ -1,30 +1,0 @@
-package com.bdool.memberhubservice.notification.domain.setting.controller;
-
-import com.bdool.memberhubservice.notification.domain.setting.entity.NotificationTargetSetting;
-import com.bdool.memberhubservice.notification.domain.setting.entity.NotificationTargetType;
-import com.bdool.memberhubservice.notification.domain.setting.entity.model.NotificationTargetSettingModel;
-import com.bdool.memberhubservice.notification.domain.setting.service.NotificationTargetSettingService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/api/notification-settings")
-@RequiredArgsConstructor
-public class NotificationTargetSettingController {
-
-    private final NotificationTargetSettingService targetSettingService;
-
-    @PostMapping("/update")
-    public ResponseEntity<NotificationTargetSetting> updateTargetNotificationSetting(@RequestBody NotificationTargetSettingModel model) {
-        return ResponseEntity.ok(targetSettingService.updateTargetNotificationSetting(model));
-    }
-
-    @GetMapping("/check")
-    public ResponseEntity<Boolean> isNotificationEnabledForTarget(
-            @RequestParam Long profileId,
-            @RequestParam Long targetId,
-            @RequestParam NotificationTargetType targetType) {
-        return ResponseEntity.ok(targetSettingService.isNotificationEnabledForTarget(profileId, targetId, targetType));
-    }
-}
