+++ notification-service/src/main/java/com/bdool/notificationservice/NotificationServiceApplication.java
@@ -1,0 +1,13 @@
+package com.bdool.notificationservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class NotificationServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(NotificationServiceApplication.class, args);
+    }
+
+}
