+++ notification-service/src/test/java/com/bdool/notificationservice/NotificationServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.bdool.notificationservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NotificationServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
